+++ lib/source_impl.cc
@@ -309,9 +309,8 @@ void source_impl::set_oversampling(int oversample) {
     device_handler::getInstance().set_oversampling(stored.device_number, oversample);
 }
 
-void source_impl::set_tcxo_dac(int dacVal) {
+void source_impl::set_tcxo_dac(uint16_t dacVal) {
     device_handler::getInstance().set_tcxo_dac(stored.device_number, dacVal);
-    //return dacVal;
 }
 
 } // namespace limesdr
